Make nose-paremetrised optional for theano.sparse
