+++ theano/sparse/basic.py
@@ -157,87 +157,6 @@ def as_sparse_or_tensor_variable(x, name=None):
         return theano.tensor.as_tensor_variable(x, name)
 
 
-def verify_grad_sparse(op, pt, structured=False, *args, **kwargs):
-    """
-    Wrapper for theano.test.unittest_tools.py:verify_grad wich
-    converts sparse variables back and forth.
-
-    Parameters
-    ----------
-    op
-        Op to check.
-    pt
-        List of inputs to realize the tests.
-    structured
-        True to tests with a structured grad, False otherwise.
-    args
-        Other `verify_grad` parameters if any.
-    kwargs
-        Other `verify_grad` keywords if any.
-
-    Returns
-    -------
-    None
-
-    """
-
-    def conv_none(x):
-        return x
-
-    def conv_csr(ind, indptr, shp):
-        def f(spdata):
-            return CSR(spdata, ind, indptr, shp)
-        return f
-
-    def conv_csc(ind, indptr, shp):
-        def f(spdata):
-            return CSC(spdata, ind, indptr, shp)
-        return f
-
-    iconv = []
-    dpt = []
-
-    for p in pt:
-        if _is_sparse(p):
-            if structured:
-                dpt.append(p.data)
-            else:
-                dpt.append(p.toarray())
-            if p.format == 'csr':
-                if structured:
-                    iconv.append(conv_csr(p.indices[:p.size], p.indptr,
-                                          p.shape))
-                else:
-                    iconv.append(csr_from_dense)
-            elif p.format == 'csc':
-                if structured:
-                    iconv.append(conv_csc(p.indices[:p.size], p.indptr,
-                                          p.shape))
-                else:
-                    iconv.append(csc_from_dense)
-            else:
-                raise NotImplementedError("No conv for %s" % (p.format,))
-        else:
-            dpt.append(p)
-            iconv.append(conv_none)
-    output = op(*[as_sparse_or_tensor_variable(p) for p in pt])
-    if isinstance(output, (list, tuple)):
-        raise NotImplementedError("verify_grad can't deal with "
-                                  "multiple outputs")
-    if _is_sparse_variable(output):
-        oconv = DenseFromSparse(structured=structured)
-    else:
-        oconv = conv_none
-
-    def conv_op(*inputs):
-        ipt = [conv(i) for i, conv in zip(inputs, iconv)]
-        out = op(*ipt)
-        return oconv(out)
-
-    return utt.verify_grad(conv_op, dpt, *args, **kwargs)
-verify_grad_sparse.E_grad = utt.verify_grad.E_grad
-
-
 def constant(x, name=None):
     if not isinstance(x, scipy.sparse.spmatrix):
         raise TypeError("sparse.constant must be called on a "
