+++ theano/sparse/sandbox/test_sp.py
@@ -20,7 +20,6 @@ from theano.compat import next
 from theano.sparse.sandbox import sp
 from theano.sparse.tests.test_basic import random_lil
 from theano.tests import unittest_tools as utt
-from theano.sparse import verify_grad_sparse
 from theano.sparse.tests.test_basic import sparse_random_inputs
 from theano.tests.unittest_tools import attr
